kkkk

+++ servicee/src/main/java/com/guigu/service/impl/ProcedureListServiceImpl.java
@@ -4,11 +4,15 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.guigu.mapper.DesignProcedureMapper;
 import com.guigu.mapper.ProcedureListMapper;
+import com.guigu.pojo.DesignProcedureModule;
+import com.guigu.pojo.Module;
 import com.guigu.pojo.ProcedureList;
 import com.guigu.service.DesignProcedureDetailsService;
+import com.guigu.service.DesignProcedureModuleService;
 import com.guigu.service.ProcedureListService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -20,9 +24,19 @@ public class ProcedureListServiceImpl extends ServiceImpl<ProcedureListMapper, P
     @Autowired
     DesignProcedureDetailsService designProcedureDetailsService;
 
+    @Autowired
+    DesignProcedureModuleService designProcedureModuleService;
+
+    @Transactional
     @Override
     public boolean delprolist(Integer proid, Integer id) {
-        designProcedureMapper.upddestag(id);
-        return designProcedureDetailsService.removeById(proid);
+        QueryWrapper<DesignProcedureModule> wrapper = new QueryWrapper<DesignProcedureModule>();
+        wrapper.eq("PRODUCT_ID", id);
+        List<DesignProcedureModule> list = designProcedureModuleService.list(wrapper);
+        if (list.size()>0) {
+            designProcedureMapper.upddestag(id);
+            return designProcedureDetailsService.removeById(proid);
+        }
+        return false;
     }
 }
